update page nav
